requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-JWT-Extended==4.6.0

# Database
psycopg2-binary==2.9.9
alembic==1.13.1

# Stripe Payment
stripe==7.8.0

# Redis & Celery (background tasks)
redis==5.0.1
celery==5.3.4

# Email
python-dotenv==1.0.0

# Validation
email-validator==2.1.0
phonenumbers==8.13.26

# VAT Validation
zeep==4.2.1  # для SOAP запросов к EU VIES

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Security
bcrypt==4.1.2
werkzeug==3.0.1

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0
